Move bitmap glyph handling into its own module
`text/flatten.rs` is about walking the positioned glyphs of a span, and
it had grown a bitmap decoder, a PNG encoder and the strike lookup on
top of that. Those move to `text/bitmap.rs`, together with their tests.

The raster cache key becomes a struct with a documented field per thing
a bitmap glyph depends on, rather than a tuple, and doubles as the
argument of the lookup so the two cannot disagree.
